pyproject.toml
[tool.poetry]
name = "data-loader-backend"
version = "0.1.0"
description = ""
authors = ["arthurkhe <artur.khe@osinit.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.6"
python-dotenv = "^0.21.1"
psycopg2 = "^2.9.5"
kafka-python = "^2.0.2"
redis = "^4.5.1"
django-timescaledb = "^0.2.13"
pytz = "^2022.7.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
